(pure-move) migrate is_uniform_write
