+++ tt_metal/impl/tensor/tensor_apis.cpp
@@ -1,0 +1,27 @@
+// SPDX-FileCopyrightText: © 2026 Tenstorrent USA, Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include <tt-metalium/experimental/tensor/tensor_apis.hpp>
+#include <tt-metalium/experimental/tensor/impl/tensor_impl.hpp>
+
+namespace tt::tt_metal {
+
+// ======================================================================================
+//                            Transfer classification
+// ======================================================================================
+
+bool is_uniform_write(const HostTensor& host_tensor, const distributed::MeshDevice& device) {
+    const auto& device_mesh_shape = device.shape();
+    const auto& host_buffer = host_tensor.buffer();
+
+    if (host_buffer.shape() != device_mesh_shape) {
+        return false;
+    }
+
+    auto all_coords = distributed::MeshCoordinateRange(device_mesh_shape);
+    return std::ranges::all_of(
+        all_coords, [&](const auto& coord) { return host_buffer.shard_coords().contains(coord); });
+}
+
+}  // namespace tt::tt_metal
