+++ ttnn/api/ttnn/tensor/tensor_impl.hpp
@@ -64,32 +64,6 @@ MeshTensor allocate_mesh_tensor(
     const TensorSpec& tensor_spec, distributed::MeshDevice& device, TensorTopology topology);
 
 // ======================================================================================
-//                        Transfer classification
-// ======================================================================================
-
-// Returns true if a H2D between the HostTensor and target MeshDevice is uniform.
-// A transfer is uniform if the host shards cover the entire shape of the MeshDevice.
-//
-// Example of uniform transfer:
-// HostTensor with a DistributedHostBuffer of shards [0, 0], [0, 1], [1, 0], [1, 1] (shape 2x2).
-// MeshDevice of shape 2x2.
-// Here the shards map exactly to the shape of the MeshDevice.
-//
-// Example of non-uniform transfers:
-//
-// 1: one to many replicas
-// HostTensor with a single shard at [0,0].
-// MeshDevice of shape 2x2.
-// This is a replica-based non-uniform transfer.
-//
-// 2: partial coverage:
-// HostTensor with a DistributedHostBuffer of shards [0, 0], and [1, 0]
-// MeshDevice of shape 2x2.
-// This is a partial coverage non-uniform transfer.
-// Only opposite sides of the MeshDevice will receive new data.
-bool is_uniform_write(const HostTensor& host_tensor, const distributed::MeshDevice& device);
-
-// ======================================================================================
 //                   Uniform enqueue_read/write_mesh_tensor
 // ======================================================================================
